+++ sourdough/sourdough.py
@@ -19,6 +19,7 @@ Read configuration parameters from instance EC2 or vSphere tags, then run
 chef-client.
 '''
 
+import datetime
 import json
 import logging
 import os
@@ -51,7 +52,9 @@ DEFAULT_TOML_FILE = '/etc/sourdough/sourdough.toml'
 DEFAULT_VMWARE_CONFIG = '/etc/sourdough/vmware.toml'
 DEFAULT_VSPHERE_KNOB = 'vsphere_host.toml'
 DEFAULT_WAIT_FOR_ANOTHER_CONVERGE = 600
+DISABLE_SOURDOUGH_F = "/etc/sourdough/Disable-Sourdough"
 DISABLE_VSPHERE = '/etc/sourdough/disable-vsphere'
+ENABLE_SOURDOUGH_DEBUGGING_F = "/etc/sourdough/debug-sourdough"
 
 knobsCache = {}
 vmwareTags = {}
@@ -701,10 +704,9 @@ def isDisabled():
   loadSharedLogger()
   logger = this.logger
   logger.info('Checking for disable switch')
-  disableFile = "/etc/sourdough/Disable-Sourdough"
-  logger.debug("  Checking for %s", disableFile)
-  if os.path.isfile(disableFile):
-      logger.debug("  %s found", disableFile)
+  logger.debug("  Checking for %s", DISABLE_SOURDOUGH_F)
+  if os.path.isfile(DISABLE_SOURDOUGH_F):
+      logger.debug("  %s found", DISABLE_SOURDOUGH_F)
       logger.critical('Chef converge disabled')
       return True
   logger.info('Disable switch not found')
@@ -990,6 +992,59 @@ def deregisterFromChef():
     if os.path.isfile("/etc/chef/%s" % chefFile):
       this.logger.info("  Scrubbing %s", chefFile)
       os.remove("/etc/chef/%s" % chefFile)
+
+
+def enableChefRuns():
+  '''
+  Remove the toggle file that disables Chef runs if it is present
+  '''
+
+  if not amRoot():
+    raise RuntimeError, 'This must be run as root'
+
+  if os.path.isfile(DISABLE_SOURDOUGH_F):
+    print "Removing %s" % DISABLE_SOURDOUGH_F
+    os.remove(DISABLE_SOURDOUGH_F)
+
+
+def disableChefRuns():
+  '''
+  Create the toggle file that disables Chef runs
+  '''
+
+  if not amRoot():
+    raise RuntimeError, 'This must be run as root'
+
+  with open(DISABLE_SOURDOUGH_F, 'w') as disabler:
+    print "Creating %s" % DISABLE_SOURDOUGH_F
+    disabler.write(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+
+
+def enableDebugMode():
+  '''
+  Remove the toggle file that enables Debug mode
+  '''
+  ENABLE_SOURDOUGH_DEBUGGING_F = "/etc/sourdough/debug-sourdough"
+
+  if not amRoot():
+    raise RuntimeError, 'This must be run as root'
+
+  with open(ENABLE_SOURDOUGH_DEBUGGING_F, 'w') as disabler:
+    print "Creating %s" % ENABLE_SOURDOUGH_DEBUGGING_F
+    disabler.write(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+
+
+def disableDebugMode():
+  '''
+  Create the toggle file that enables debugging mode
+  '''
+
+  if not amRoot():
+    raise RuntimeError, 'This must be run as root'
+
+  if os.path.isfile(ENABLE_SOURDOUGH_DEBUGGING_F):
+    print "Removing %s" % ENABLE_SOURDOUGH_DEBUGGING_F
+    os.remove(ENABLE_SOURDOUGH_DEBUGGING_F)
 
 
 if __name__ == '__main__':